requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
dnspython==2.7.0
eventlet==0.39.0
Faker==35.0.0
fastapi==0.115.7
Flask==3.1.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
outcome==1.3.0.post0
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
selenium==4.28.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.45.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
